pyproject.toml
[project]
name = "demhack-safety-vision-mvp"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flaml==2.2.0",
    "matplotlib>=3.9.2",
    "mergedeep>=1.3.4",
    "nltk>=3.9.1",
    "numpy==1.26.4",
    "pip>=24.2",
    "pyautogen>=0.3.0",
    "ruamel-yaml>=0.18.6",
    "ruff>=0.6.6",
    "spacy>=3.7.6",
    "telethon>=1.36.0",
]

[tool.hatch.build]
include = [
    "src"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
ignore = ['F841', 'F401', 'E203']
extend-select = ["I"]
